+++ app/javascript/block_editor/blocks/be-cover/index.js
@@ -26,11 +26,6 @@ export const settings = {
 	description:  'Add an image with a text overlay - great for breaking up content.',
   icon,
   category: 'formatting',
-	styles: [
-		{ name: 'left', label: 'Left Side', isDefault: true },
-		{ name: 'right', label: 'Right Side' },
-		{ name: 'full-width', label: 'Full Width' }
-	],
   example: {
     innerBlocks: [
       {